+++ tests/verification.spec.ts
@@ -3,7 +3,9 @@ import {describe, expect, it} from 'vitest';
 import {flattenedAddresses, ListItem} from '../ui/src/utils/getAddresses';
 import verified from './cache/verified.json';
 import {writeFileSync} from 'fs';
-import {zeroAddress} from 'viem';
+import {Hex, PublicClient, zeroAddress} from 'viem';
+import {getCode} from 'viem/actions';
+import {getClient} from '../scripts/clients';
 
 const ETHERSCAN_API_KEY = process.env.ETHERSCAN_API_KEY as string;
 
@@ -91,6 +93,22 @@ function getApiUrl(chainId: number) {
   return `https://api.etherscan.io/v2/api`;
 }
 
+const knownErrors = {
+  1: {
+    '0xD01ab9a6577E1D84F142e44D49380e23A340387d': true, // very old contracts we know will never be verified
+    '0xb51EDdDD8c47856D81C8681EA71404Cec93E92c6': true, // etherscan issue - not detecting that it's verified
+  },
+  1101: {
+    '0xF1c11BE0b4466728DDb7991A0Ac5265646ec9672': true, // very old contracts we know will never be verified
+  },
+  137: {
+    '0x645654D59A5226CBab969b1f5431aA47CBf64ab8': true, // very old contracts we know will never be verified
+  },
+  43114: {
+    '0x11979886A6dBAE27D7a72c49fCF3F23240D647bF': true, // very old contracts we know will never be verified
+  },
+};
+
 describe(
   'verification',
   () => {
@@ -98,36 +116,41 @@ describe(
       const addressesToCheck = flattenedAddresses.filter(
         (item) =>
           ![ChainId.harmony, ChainId.fantom].includes(item.chainId as any) &&
-          !ChainList[item.chainId].testnet,
+          !ChainList[item.chainId].testnet &&
+          !knownErrors[item.chainId]?.[item.value],
       );
       const errors: {item: ListItem}[] = [];
       let newVerified = false;
       // unique set of addresses checked on this iteration
       // used to prevent double checking the same address
       const checked = new Set<string>();
-      // for (const item of addressesToCheck) {
-      //   const hasBeenCheckedBefore = verified[item.chainId][item.value];
-      //   if (!hasBeenCheckedBefore && item.value !== zeroAddress) {
-      //     const key = `${item.chainId}-${item.value}`;
-      //     if (checked.has(key)) continue;
-      //     checked.add(key);
-      //     const {status, result} = (await checkVerified(item)) as {
-      //       status: string;
-      //       result: {ContractName: string}[];
-      //     };
-      //     await sleep(300);
-      //     if (status !== '1' || !result[0].ContractName) {
-      //       errors.push({item});
-      //       console.log(result);
-      //     } else {
-      //       newVerified = true;
-      //       if (!verified[item.chainId]) verified[item.chainId] = {};
-      //       verified[item.chainId][item.value] = {
-      //         name: result[0].ContractName,
-      //       };
-      //     }
-      //   }
-      // }
+      for (const item of addressesToCheck) {
+        const hasBeenCheckedBefore = verified[item.chainId]?.[item.value];
+        if (!hasBeenCheckedBefore && item.value !== zeroAddress) {
+          const key = `${item.chainId}-${item.value}`;
+          if (checked.has(key)) continue;
+          checked.add(key);
+          const client = getClient(item.chainId) as PublicClient;
+          const hasCode = await getCode(client, {address: item.value as Hex});
+          if (hasCode) {
+            const {status, result} = (await checkVerified(item)) as {
+              status: string;
+              result: {ContractName: string}[];
+            };
+            await sleep(300);
+            if (status !== '1' || !result[0].ContractName) {
+              errors.push({item});
+              console.log(item.value, result);
+            } else {
+              newVerified = true;
+              if (!verified[item.chainId]) verified[item.chainId] = {};
+              verified[item.chainId][item.value] = {
+                name: result[0].ContractName,
+              };
+            }
+          }
+        }
+      }
       if (newVerified) {
         writeFileSync('./tests/cache/verified.json', JSON.stringify(verified, null, 2), {
           encoding: 'utf-8',